requirements.txt
django==3.2.20
djangorestframework==3.14.0
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
mixer==7.2.2
celery==4.4.1
redis==3.4.1
psycopg2-binary==2.8.6 
